pyproject.toml
[build-system]
requires = []
backend-path = ["src", "tools-py/src"]
build-backend = "toolspy.project.build"

[project]
name = "tools-py"
version = "0.0.3"
description = "minimal build backend which allows to add required binaries to the project"
authors = [
  { name="Mikalai Davydzenka", email="mikalai.davydzenka@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "httpx",
  "tomlkit",
  "attrs",
  "cattrs",
  "pyyaml",
  "fire",
]

[project.scripts]
toolbox = "toolspy.toolbox:cli"

[project.urls]
Homepage = "https://github.com/toolsfab/tools-py"

# [[tool.uv.index]]
# name = "testpypi"
# url = "https://test.pypi.org/simple/"
# publish-url = "https://test.pypi.org/legacy/"
